+++ konsave/consts.py
@@ -7,6 +7,8 @@ from konsave import __version__
 
 HOME = os.path.expandvars("$HOME")
 CONFIG_DIR = os.path.join(HOME, ".config")
+SHARE_DIR = os.path.join(HOME, ".local/share")
+BIN_DIR = os.path.join(HOME, ".local/bin")
 KONSAVE_DIR = os.path.join(CONFIG_DIR, "konsave")
 PROFILES_DIR = os.path.join(KONSAVE_DIR, "profiles")
 CONFIG_FILE = os.path.join(KONSAVE_DIR, "conf.yaml")